requirements.txt
streamlit>=1.40.0,<2.0
google-genai>=1.0.0
fal-client>=0.5.0
requests>=2.31.0
pydantic>=2.5.0
eval_type_backport>=0.2.0
Pillow>=10.0.0
tenacity>=8.2.0
python-dotenv>=1.0.0
